undo touch

+++ functions.c
@@ -46,8 +46,8 @@ int print_percent(void)
  */
 int print_d(va_list arg)
 {
-  long int n = va_arg(arg, long int);
-  long int divisor = 1, count = 0;
+  int n = va_arg(arg, int);
+  int divisor = 1, count = 0;
 
   if (n < 0) {
     _putchar('-');
